Fix restart script and new appspec

+++ _scripts/ubyssey_restart.sh
@@ -1,7 +1,9 @@
 #!/bin/bash
 
 # Stop the server
-sudo kill `cat run/ubyssey.pid`
+if [ -f run/ubyssey.pid ]; then
+  sudo kill `cat run/ubyssey.pid`
+fi
 
 # Start the server
 sudo -u ubyssey -H sh -c "_scripts/gunicorn_start"
